requirements.txt
streamlit
speechrecognition
edge-tts
openai
google-generativeai